Prevent infinite loops in octarine deserialisation
The jackson JsonParser nextToken() method will always return null once the end of the input has been reached. We therefore need to check the parser state in order to prevent infinite loops.

We also add an additional check to ListDeserialiser when called from the top level which makes it more robust when parsing JSON with a list instead of an object as the root element.