Prevent null entries in JSON object from causing other values to be omitted
There was a bug in the handling of nested records which meant that a null value for a field would cause the record deserialiser to lose values after the null. This change adds a check for this situation and handles it correctly.

Have also added a test case to verify the correct behaviour.